pyproject.toml
[project]
name = "pinchana-core"
version = "0.1.0"
description = "Shared models, storage, and VPN controller for scraper services"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.27.0",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.2",
    "tenacity>=9.1.4",
]

[build-system]
requires = ["uv_build>=0.11.8,<0.12.0"]
build-backend = "uv_build"
